src/backend/requirements.txt
flask
flask-cors
flask-jwt-extended
google-generativeai
pymongo 
bcrypt 
python-dotenv 
smtplib